Improve naming of the elements. Remove unnecessary elements. Implement some nit suggestions from the reviewer
