--- lib/basic.ex
@@ -1,18 +1,0 @@
-defmodule Basic do
-  @moduledoc """
-  Documentation for `Basic`.
-  """
-
-  @doc """
-  Hello world.
-
-  ## Examples
-
-      iex> Basic.hello()
-      :world
-
-  """
-  def hello do
-    :world
-  end
-end
